Broken! Used intellij to copy over files from original project.
Committing in broken state so that it's in version control for comparison reference.
Will need to make this project depend on java.bean, and do extra finessing to get
this to fit for tripleplay/playn/android/non swing.

Decided not to fork original because I care more about the playn project structure
to be correct, so I don't have to fiddle.